+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 #### master
 
+- CHANGED: Renamed Record struct to ZoneRecord (GH-18).
+
 - CHANGED: Setting a custom user-agent no longer overrides the origina user-agent (GH-15)
 
 - REMOVED: Removed support for wildcard accounts (GH-16).